+++ packages/maplibre-react-components/src/hooks/useControl.ts
@@ -23,20 +23,16 @@ export function useControl({
   const onRemoveStable = useEventCallback(onRemove || null);
 
   useEffect(() => {
-    console.log("useControl useEffect");
-
     if (!map.hasControl(ctrl)) {
-      console.log("map addControl");
       map.addControl(ctrl, position);
     }
     return () => {
       onRemoveStable && onRemoveStable(map);
 
       if (map.hasControl(ctrl)) {
-        console.log("map removeControl uhh");
         map.removeControl(ctrl);
       } else {
-        console.log("map has not control", ctrl);
+        console.error("map has not control", ctrl);
       }
     };
   }, [map, ctrl, onRemoveStable, position]);