+++ packages/pentatrion-design/components/button/index.ts
@@ -1,2 +1,3 @@
 export { Button, type ButtonProps, buttonVariants } from "./Button";
 export { ButtonGroup } from "./ButtonGroup";
+export { LinkButton } from "./LinkButton";